fix(homepage): ambiguous conversion

+++ src/Transmitron/Tabs/Homepage.cpp
@@ -279,9 +279,9 @@ void Homepage::fillPropertyGrid(
   pfp.at(Properties::AutoReconnect)->SetValue(brokerOptions.getAutoReconnect());
   pfp.at(Properties::ClientId)->SetValue(brokerOptions.getClientId());
   pfp.at(Properties::ConnectTimeout)->SetValue((int)brokerOptions.getConnectTimeout().count());
-  pfp.at(Properties::DisconnectTimeout)->SetValue(brokerOptions.getDisconnectTimeout().count());
+  pfp.at(Properties::DisconnectTimeout)->SetValue((int)brokerOptions.getDisconnectTimeout().count());
   pfp.at(Properties::Hostname)->SetValue(brokerOptions.getHostname());
-  pfp.at(Properties::KeepAlive)->SetValue(brokerOptions.getKeepAliveInterval().count());
+  pfp.at(Properties::KeepAlive)->SetValue((int)brokerOptions.getKeepAliveInterval().count());
   pfp.at(Properties::MaxInFlight)->SetValue((int)brokerOptions.getMaxInFlight());
   pfp.at(Properties::MaxReconnectRetries)->SetValue((int)brokerOptions.getMaxReconnectRetries());
   pfp.at(Properties::Name)->SetValue(name);
